Add ambient heroes for philosophy and site map
